requirements.txt
accelerate==0.27.2
aiohttp==3.11.9
anthropic==0.46.0
auto_gptq==0.7.1
dill==0.3.1.1
evaluate==0.4.1
hf_transfer==0.1.9
huggingface_hub==0.27.0
immutabledict==4.2.1
Jinja2==3.1.3
jsonlines==4.0.0
langdetect==1.0.9
loguru==0.7.2
mamba_ssm==2.2.4
matplotlib==3.8.3
more_itertools==10.6.0
nltk==3.8.1
numpy==2.2.3
openai==0.28.0
optimum==1.24.0
packaging==24.2
pandas==2.2.3
peft==0.9.0
promptsource==0.2.3
psutil==6.1.0
pycountry==24.6.1
pytablewriter==1.2.0
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
ray==2.40.0
regex==2023.12.25
Requests==2.32.3
rouge_score==0.1.2
sacrebleu==2.4.0
scikit_learn==1.4.1.post1
scipy==1.15.2
seaborn==0.13.2
sentence_transformers==3.0.1
setuptools==68.2.2
sqlitedict==2.1.0
sympy==1.12
thefuzz==0.22.1
tiktoken==0.6.0
torch==2.2.1
tqdm==4.66.2
tqdm_multiprocess==0.0.11
transformers==4.47.1
vllm==0.3.3
word2number==1.1
zeno_client==0.1.16
zstandard==0.22.0
